refactor: move VC to domain chore: remove core folder
