+++ ProtoCredentials/OpenID4VC-Prototype/Application/Interfaces/IVerifierService.cs
@@ -1,4 +1,5 @@
 ﻿using OpenID4VC_Prototype.Application.Models;
+using OpenID4VC_Prototype.Domain.Models;
 
 namespace OpenID4VC_Prototype.Application.Interfaces;
