+++ ProtoCredentials/OpenID4VC-Prototype/Application/Services/VerifierService.cs
@@ -1,6 +1,7 @@
 ﻿using OpenID4VC_Prototype.Application.Interfaces;
 using OpenID4VC_Prototype.Application.Models;
 using OpenID4VC_Prototype.Domain.Interfaces;
+using OpenID4VC_Prototype.Domain.Models;
 using OpenID4VC_Prototype.Utils;
 
 namespace OpenID4VC_Prototype.Application.Services;
