+++ ProtoCredentials/OpenID4VC-Prototype/Domain/Interfaces/ICryptoService.cs
@@ -1,4 +1,6 @@
-﻿namespace OpenID4VC_Prototype.Domain.Interfaces;
+﻿using OpenID4VC_Prototype.Domain.Models;
+
+namespace OpenID4VC_Prototype.Domain.Interfaces;
 
 public interface ICryptoService
 {
